# startUp/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AiEd Learning Assistant",
    "version": "1.0.0",
    "description": "AiEd Learning Assistant API Documentation"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "lessons": {
        "type": "object",
        "required": ["title", "start_date", "memorial_script", "id"],
        "properties": {
          "title": {
            "type": "string"
          },
          "start_date": {
            "type": "string"
          },
          "memorial_script": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "cover": {
            "type": "string",
            "nullable": true
          },
          "end_date": {
            "type": "string",
            "nullable": true
          },
          "index": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/{lang}/quarters/{quarter_id}/lessons": {
      "post": {
        "summary": "Create Lesson",
        "description": "Create Lesson",
        "tags": ["Lessons"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quarter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/lessons"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "summary": "Get all Lessons",
        "description": "Get all Lessons by language and quarter",
        "tags": ["Lessons"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "en"
          },
          {
            "name": "quarter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2023_03"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/lessons"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/{lang}/quarters/{quarter_id}/lessons/{lesson_id}": {
      "get": {
        "summary": "Get Specific Lesson",
        "description": "Get Specific Lesson",
        "tags": ["Lessons"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "en"
          },
          {
            "name": "quarter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2023_03"
          },
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "01"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/lessons"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Lesson",
        "description": "Update Lesson",
        "tags": ["Lessons"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quarter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/lessons"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/lessons"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Lesson",
        "description": "Delete Lesson",
        "tags": ["Lessons"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quarter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  }
}